+++ ReleaseNotes/release_notes.md
@@ -2,7 +2,8 @@
 ## New features
 
  1. minor tweak for the treatment of `delta` for `Ostap::Math::Sigmoid`. It should improve the fits.
- 
+ 1. many fixes and tweaks  for `DNN` & `USTAT` Goodness of fit estimators 
+
 ## Bug fixes
 
 ## Backward incompatible